Move `Collider` into `collision` module
